pyproject.toml
[build-system]
requires = ["hatchling>=1.27"]
build-backend = "hatchling.build"

[project]
name = "ai-investing"
version = "0.1.0"
description = "Config-driven multi-agent investment research platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Ethan Shin" }
]
dependencies = [
  "fastapi>=0.116,<1.0",
  "httpx>=0.28,<0.29",
  "langchain-core>=0.3,<0.4",
  "langgraph>=0.6,<0.7",
  "pydantic>=2.11,<3.0",
  "pydantic-settings>=2.10,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "python-dateutil>=2.9,<3.0",
  "pyyaml>=6.0,<7.0",
  "sqlalchemy>=2.0,<3.0",
  "typer>=0.16,<0.17",
  "uvicorn[standard]>=0.35,<0.36"
]

[project.optional-dependencies]
anthropic = [
  "langchain-anthropic>=0.3,<0.4"
]
openai = [
  "langchain-openai>=0.3,<0.4"
]
dev = [
  "mypy>=1.17,<2.0",
  "pytest>=8.4,<9.0",
  "pytest-cov>=6.2,<7.0",
  "ruff>=0.12,<0.13"
]

[project.scripts]
ai-investing = "ai_investing.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_investing"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "S", "C4"]
ignore = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
packages = ["ai_investing"]
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict = true

